LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/reattach-container-with-dirty-child-expected.txt

XXX

FAIL Initially wider than 200px assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
FAIL Container query changed and inner.style applied assert_equals: expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
